scripts/requirements.txt
gitpython==3.1.24
requests==2.26.0
PyGithub